fix: Move settings code to `Settings` Class instead of `allTimeData` Class
