+++ packages/tosu/src/entities/AllTimesData/index.ts
@@ -1,510 +1,36 @@
 import { wLogger } from '@tosu/common';
-import { Process } from 'tsprocess/dist/process';
-
-import { Bindings, VirtualKeyCode } from '@/utils/bindings';
 
 import { AbstractEntity } from '../AbstractEntity';
-import { Settings } from '../Settings';
-
-interface IBindable {
-    setValue: (settings: Settings, value: any) => void;
-}
-
-interface IConfigBindable extends IBindable {
-    type: 'bool' | 'byte' | 'int' | 'double' | 'string' | 'bstring' | 'enum';
-}
 
 export class AllTimesData extends AbstractEntity {
+    IsWatchingReplay: number = 0;
+    ShowInterface: boolean = false;
+
+    ChatStatus: number = 0;
     Status: number = 0;
+
     GameTime: number = 0;
     PlayTime: number = 0;
     MenuMods: number = 0;
-    ChatStatus: number = 0;
+
+    GameFolder: string = '';
     SkinFolder: string = '';
     SongsFolder: string = '';
-    ShowInterface: boolean = false;
-    IsWatchingReplay: number = 0;
+    MemorySongsFolder: string = '';
 
-    private configStateErrorAttempts: number = 0;
-    private bindingStateErrorAttempts: number = 0;
-
-    private configList: Record<string, IConfigBindable> = {
-        VolumeUniversal: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.audio.volume.master = value;
-            }
-        },
-        VolumeEffect: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.audio.volume.effect = value;
-            }
-        },
-        VolumeMusic: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.audio.volume.music = value;
-            }
-        },
-        _ReleaseStream: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.client.branch = value;
-            }
-        },
-        DimLevel: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.background.dim = value;
-            }
-        },
-        ShowStoryboard: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.background.storyboard = value;
-            }
-        },
-        ShowInterface: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.showInterface = value;
-            }
-        },
-        BeatmapDirectory: {
-            type: 'bstring',
-            setValue: (settings, value) => {
-                settings.BeatmapDirectory = value;
-            }
-        },
-        ScoreMeter: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.scoreMeter.type = value;
-            }
-        },
-        ScoreMeterScale: {
-            type: 'double',
-            setValue: (settings, value) => {
-                settings.scoreMeter.size = parseFloat((value || 0).toFixed(2));
-            }
-        },
-        Offset: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.offset.universal = value;
-            }
-        },
-        CursorSize: {
-            type: 'double',
-            setValue: (settings, value) => {
-                settings.cursor.size = parseFloat((value || 0).toFixed(2));
-            }
-        },
-        MouseSpeed: {
-            type: 'double',
-            setValue: (settings, value) => {
-                settings.mouse.sensitivity = parseFloat(
-                    (value || 0).toFixed(2)
-                );
-            }
-        },
-        Fullscreen: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.resolution.fullscreen = value;
-            }
-        },
-        Width: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.resolution.width = value;
-            }
-        },
-        Height: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.resolution.height = value;
-            }
-        },
-        WidthFullscreen: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.resolution.widthFullscreen = value;
-            }
-        },
-        HeightFullscreen: {
-            type: 'int',
-            setValue: (settings, value) => {
-                settings.resolution.heightFullscreen = value;
-            }
-        },
-        AutomaticCursorSizing: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.cursor.autoSize = value;
-            }
-        },
-        IgnoreBeatmapSamples: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.audio.ignoreBeatmapSounds = value;
-            }
-        },
-        SkinSamples: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.audio.useSkinSamples = value;
-            }
-        },
-        LastVersion: {
-            type: 'bstring',
-            setValue: (settings, value) => {
-                settings.client.version = value;
-            }
-        },
-        ManiaSpeedBPMScale: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.mania.speedBPMScale = value;
-            }
-        },
-        UsePerBeatmapManiaSpeed: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.mania.usePerBeatmapSpeedScale = value;
-            }
-        },
-        MouseDisableButtons: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.mouse.disableButtons = value;
-            }
-        },
-        MouseDisableWheel: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.mouse.disableWheel = value;
-            }
-        },
-        ProgressBarType: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.progressBarType = value;
-            }
-        },
-        RankType: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.leaderboardType = value;
-            }
-        },
-        UpdatePending: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.client.updateAvailable = value;
-            }
-        },
-
-        UseSkinCursor: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.cursor.useSkinCursor = value;
-            }
-        },
-        RawInput: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.mouse.rawInput = value;
-            }
-        },
-        TreeSortMode: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.groupType = value;
-            }
-        },
-        TreeSortMode2: {
-            type: 'enum',
-            setValue: (settings, value) => {
-                settings.sortType = value;
-            }
-        },
-        EditorDefaultSkin: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.skin.useDefaultSkinInEditor = value;
-            }
-        },
-        ComboColourSliderBall: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.skin.tintSliderBall = value;
-            }
-        },
-        IgnoreBeatmapSkins: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.skin.ignoreBeatmapSkins = value;
-            }
-        },
-        Skin: {
-            type: 'bstring',
-            setValue: (settings, value) => {
-                settings.skin.name = value;
-            }
-        },
-        UseTaikoSkin: {
-            type: 'bool',
-            setValue: (settings, value) => {
-                settings.skin.useTaikoSkin = value;
-            }
-        }
-    };
-
-    private bindingList: Record<number, IBindable> = {
-        [Bindings.OsuLeft]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.osu.k1 = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.OsuRight]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.osu.k2 = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.OsuSmoke]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.osu.smokeKey = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.FruitsDash]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.fruits.Dash = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.FruitsLeft]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.fruits.k1 = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.FruitsRight]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.fruits.k2 = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.TaikoInnerLeft]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.taiko.innerLeft = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.TaikoInnerRight]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.taiko.innerRight = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.TaikoOuterLeft]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.taiko.outerLeft = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.TaikoOuterRight]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.taiko.outerRight = VirtualKeyCode[value];
-            }
-        },
-        [Bindings.QuickRetry]: {
-            setValue: (settings, value: number) => {
-                settings.keybinds.quickRetry = VirtualKeyCode[value];
-            }
-        }
-    };
-
-    setConfigValue(
-        process: Process,
-        settings: Settings,
-        address: number,
-        position: number = 0
-    ) {
-        try {
-            const current =
-                process.readInt(address + 0x8) + 0x8 + 0x10 * position;
-            const keyAddress = process.readInt(current);
-
-            const key = process.readSharpString(keyAddress);
-            const bindable = process.readInt(current + 0x4);
-
-            let value: any;
-            switch (this.configList[key].type) {
-                case 'byte':
-                    value = process.readByte(bindable + 0xc);
-                    break;
-                case 'bool':
-                    value = Boolean(process.readByte(bindable + 0xc));
-                    break;
-                case 'int':
-                case 'double':
-                    value = process.readDouble(bindable + 0x4);
-                    break;
-                case 'string':
-                    value = process.readSharpString(
-                        process.readInt(current + 0x4)
-                    );
-                    break;
-                case 'bstring':
-                    value = process.readSharpString(
-                        process.readInt(bindable + 0x4)
-                    );
-                    break;
-                case 'enum':
-                    value = process.readInt(bindable + 0xc);
-                    break;
-                default:
-                    break;
-            }
-
-            if (value != null) {
-                // console.log(position, key, value);
-
-                this.configList[key].setValue(settings, value);
-            }
-        } catch (exc) {
-            wLogger.error("ATD(setConfigValue) Can't update config state");
-            wLogger.debug(exc);
-        }
+    setGameFolder(value: string) {
+        this.GameFolder = value;
     }
 
-    setBindingValue(
-        process: Process,
-        settings: Settings,
-        address: number,
-        position: number = 0
-    ) {
-        try {
-            const current =
-                process.readInt(address + 0x8) + 0x8 + 0x10 * position;
-
-            const key = process.readInt(current);
-            const value = process.readInt(current + 0xc);
-
-            const bindable = this.bindingList[key];
-            if (bindable) {
-                // console.log(position, Bindings[key], VirtualKeyCode[value]);
-
-                bindable.setValue(settings, value);
-            }
-        } catch (exc) {
-            wLogger.error("ATD(setBindingValue) Can't update config state");
-            wLogger.debug(exc);
-        }
-    }
-
-    // preventSpamArray: (number | string)[] = [];
-
-    updateConfigState(
-        process: Process,
-        settings: Settings,
-        configurationAddr: number
-    ) {
-        try {
-            const values = [
-                0, 18, 20, 21, 43, 44, 45, 53, 67, 68, 74, 78, 88, 89, 91, 92,
-                93, 94, 95, 100, 101, 102, 104, 110, 111, 116, 124, 125, 126,
-                130, 132, 134, 140, 144, 149, 158, 159, 233
-            ];
-            for (const position of values) {
-                this.setConfigValue(
-                    process,
-                    settings,
-                    configurationAddr,
-                    position
-                );
-            }
-
-            // // KEEP AS THE REFERENCE TO POSITION OF THE VALUES
-            // const rawSharpDictionary = process.readSharpDictionary(configurationAddr);
-            // for (let i = 0; i < rawSharpDictionary.length; i++) {
-            //     const current = rawSharpDictionary[i];
-            //     const keyAddress = process.readInt(current);
-
-            //     const key = process.readSharpString(keyAddress);
-
-            //     if (key in this.configList || this.preventSpamArray.includes(key)) {
-            //         continue;
-            //     }
-
-            //     console.log(i, current, key);
-            //     this.preventSpamArray.push(key);
-            // }
-
-            if (this.configStateErrorAttempts !== 0) {
-                this.configStateErrorAttempts = 0;
-            }
-        } catch (exc) {
-            this.configStateErrorAttempts += 1;
-
-            if (this.configStateErrorAttempts > 5) {
-                wLogger.error(
-                    "ATD(updateConfigState) Can't update config state"
-                );
-            }
-            wLogger.debug(exc);
-        }
-    }
-
-    updateBindingState(
-        process: Process,
-        settings: Settings,
-        bindingConfigAddr: number
-    ) {
-        try {
-            const values = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 18];
-            for (const position of values) {
-                this.setBindingValue(
-                    process,
-                    settings,
-                    bindingConfigAddr,
-                    position
-                );
-            }
-
-            // // KEEP AS THE REFERENCE TO POSITION OF VALUES
-            // const rawSharpDictionary = process.readSharpDictionary(bindingConfigAddr);
-            // for (let i = 0; i < rawSharpDictionary.length; i++) {
-            //     const current = rawSharpDictionary[i];
-            //     const key = process.readInt(current);
-            //     const value = process.readInt(current + 0xc);
-
-            //     if (key in this.bindingList || this.preventSpamArray.includes(key)) {
-            //         continue;
-            //     }
-            //     const bindable = Bindings[key];
-            //     console.log(i, current, bindable, key, value);
-
-            //     this.preventSpamArray.push(key);
-            // }
-
-            if (this.bindingStateErrorAttempts !== 0) {
-                this.bindingStateErrorAttempts = 0;
-            }
-        } catch (exc) {
-            this.bindingStateErrorAttempts += 1;
-
-            if (this.bindingStateErrorAttempts > 5) {
-                wLogger.error(
-                    "ATD(updateBindingState) Can't update binding state"
-                );
-            }
-            wLogger.debug(exc);
-        }
+    setSongsFolder(value: string) {
+        this.SongsFolder = value;
     }
 
     updateState() {
         try {
-            const { process, patterns, settings } = this.services.getServices([
+            const { process, patterns } = this.services.getServices([
                 'process',
-                'patterns',
-                'settings'
+                'patterns'
             ]);
 
             const {
@@ -514,9 +40,7 @@ export class AllTimesData extends AbstractEntity {
                 chatCheckerAddr,
                 skinDataAddr,
                 settingsClassAddr,
-                canRunSlowlyAddr,
-                configurationAddr,
-                bindingsAddr
+                canRunSlowlyAddr
                 // gameTimePtr,
             } = patterns.getPatterns([
                 'statusPtr',
@@ -525,9 +49,7 @@ export class AllTimesData extends AbstractEntity {
                 'chatCheckerAddr',
                 'skinDataAddr',
                 'settingsClassAddr',
-                'canRunSlowlyAddr',
-                'configurationAddr',
-                'bindingsAddr'
+                'canRunSlowlyAddr'
                 // 'gameTimePtr',
             ]);
 
@@ -554,7 +76,7 @@ export class AllTimesData extends AbstractEntity {
             this.IsWatchingReplay = process.readByte(
                 process.readInt(canRunSlowlyAddr + 0x46)
             );
-            this.SongsFolder = process.readSharpString(
+            this.MemorySongsFolder = process.readSharpString(
                 process.readInt(
                     process.readInt(
                         process.readInt(settingsClassAddr + 0x8) + 0xb8
@@ -569,18 +91,6 @@ export class AllTimesData extends AbstractEntity {
                         process.readInt(settingsClassAddr + 0x8) + 0x4
                     ) + 0xc
                 )
-            );
-
-            this.updateConfigState(
-                process,
-                settings,
-                process.readPointer(configurationAddr)
-            );
-
-            this.updateBindingState(
-                process,
-                settings,
-                process.readPointer(bindingsAddr)
             );
         } catch (exc) {
             wLogger.error(`ATD(updateState) ${(exc as any).message}`);
